requirements.txt
gradio
sentence-transformers
faiss-cpu
PyMuPDF
ollama
